+++ frontend/app/FooterContent.tsx
@@ -117,10 +117,12 @@ export default function FooterContent() {
           padding: "0.3rem 0.75rem",
           borderRadius: "0.4rem",
           textDecoration: "none",
+          whiteSpace: "nowrap",
         }}
       >
-        <svg width="13" height="13" viewBox="0 0 24 24" fill="currentColor" aria-hidden="true" focusable="false">
-          <path d="M12 0C5.37 0 0 5.37 0 12c0 5.3 3.438 9.8 8.205 11.385.6.113.82-.258.82-.577 0-.285-.01-1.04-.015-2.04-3.338.724-4.042-1.61-4.042-1.61-.546-1.385-1.335-1.755-1.335-1.755-1.087-.744.084-.729.084-.729 1.205.084 1.838 1.236 1.838 1.236 1.07 1.835 2.809 1.305 3.495.998.108-.776.417-1.305.76-1.605-2.665-.3-5.466-1.332-5.466-5.93 0-1.31.465-2.38 1.235-3.22-.135-.303-.54-1.523.105-3.176 0 0 1.005-.322 3.3 1.23.96-.267 1.98-.399 3-.405 1.02.006 2.04.138 3 .405 2.28-1.552 3.285-1.23 3.285-1.23.645 1.653.24 2.873.12 3.176.765.84 1.23 1.91 1.23 3.22 0 4.61-2.805 5.625-5.475 5.92.42.36.81 1.096.81 2.22 0 1.606-.015 2.896-.015 3.286 0 .315.21.69.825.57C20.565 21.795 24 17.295 24 12c0-6.63-5.37-12-12-12z"/>
+        {/* GitHub Octocat logo */}
+        <svg width="14" height="14" viewBox="0 0 98 96" fill="currentColor" aria-hidden="true" focusable="false">
+          <path fillRule="evenodd" clipRule="evenodd" d="M48.854 0C21.839 0 0 22 0 49.217c0 21.756 13.993 40.172 33.405 46.69 2.427.49 3.316-1.059 3.316-2.362 0-1.141-.08-5.052-.08-9.127-13.59 2.934-16.42-5.867-16.42-5.867-2.184-5.704-5.42-7.17-5.42-7.17-4.448-3.015.324-3.015.324-3.015 4.934.326 7.523 5.052 7.523 5.052 4.367 7.496 11.404 5.378 14.235 4.074.404-3.178 1.699-5.378 3.074-6.6-10.839-1.141-22.243-5.378-22.243-24.283 0-5.378 1.94-9.778 5.014-13.2-.485-1.222-2.184-6.275.486-13.038 0 0 4.125-1.304 13.426 5.052a46.97 46.97 0 0 1 12.214-1.63c4.125 0 8.33.571 12.213 1.63 9.302-6.356 13.427-5.052 13.427-5.052 2.67 6.763.97 11.816.485 13.038 3.155 3.422 5.015 7.822 5.015 13.2 0 18.905-11.404 23.06-22.324 24.283 1.78 1.548 3.316 4.481 3.316 9.126 0 6.6-.08 11.897-.08 13.526 0 1.304.89 2.853 3.316 2.364 19.412-6.52 33.405-24.935 33.405-46.691C97.707 22 75.788 0 48.854 0z"/>
         </svg>
         GitHub
       </a>
@@ -157,20 +159,21 @@ export default function FooterContent() {
         style={{
           display: "inline-flex",
           alignItems: "center",
-          gap: "0.35rem",
-          backgroundColor: "#0C2461",
-          color: "#fff",
+          gap: "0.4rem",
+          backgroundColor: "#FFCC00",
+          color: "#000",
           fontSize: "0.7rem",
-          fontWeight: 700,
+          fontWeight: 800,
           padding: "0.3rem 0.75rem",
           borderRadius: "0.4rem",
           textDecoration: "none",
-          letterSpacing: "0.04em",
+          letterSpacing: "0.06em",
+          whiteSpace: "nowrap",
         }}
       >
-        {/* TWINT brand icon: stylised phone with payment wave */}
+        {/* TWINT T-Lettermark */}
         <svg width="13" height="13" viewBox="0 0 24 24" fill="currentColor" aria-hidden="true" focusable="false">
-          <path d="M17 1H7C5.34 1 4 2.34 4 4v16c0 1.66 1.34 3 3 3h10c1.66 0 3-1.34 3-3V4c0-1.66-1.34-3-3-3zm-5 20c-.83 0-1.5-.67-1.5-1.5S11.17 18 12 18s1.5.67 1.5 1.5S12.83 21 12 21zm5-4H7V4h10v13z"/>
+          <path d="M2 3h20v4.5h-8v13.5h-4V7.5H2z"/>
         </svg>
         TWINT
       </a>
@@ -188,15 +191,14 @@ export default function FooterContent() {
           color: "#fff",
           fontSize: "0.7rem",
           fontWeight: 600,
-          padding: "0.3rem 0.75rem",
+          padding: "0.3rem 0.9rem",
           borderRadius: "0.4rem",
           textDecoration: "none",
+          whiteSpace: "nowrap",
         }}
       >
         {/* eslint-disable-next-line @next/next/no-img-element */}
-        {/* eslint-disable-next-line @next/next/no-img-element */}
-        {/* eslint-disable-next-line @next/next/no-img-element */}
-        <img src="/transivroom-icon.png" alt="" width={16} height={16} style={{ filter: "brightness(0) invert(1) brightness(2)" }} />
+        <img src="/transivroom-icon.png" alt="" width={14} height={14} style={{ filter: "brightness(0) invert(1) brightness(2)", flexShrink: 0 }} />
         transivroom
       </a>